Display notes for App CRs
The displayed content is .status.release.reason

The status of an App CR is synced over from the associated Chart CR.
The reason field contains information on the outcome of the last Helm
release, only for error scenarios. The other fields reflect the actual
current state of the Helm release.